# backend/cmd/ingest/docs/swagger.yaml
basePath: /ingest_v1
definitions:
  routes.OpenScraperAttachment:
    properties:
      document_type:
        type: string
      extra_metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      url:
        type: string
    type: object
  routes.OpenScraperFiling:
    properties:
      attachments:
        items:
          $ref: '#/definitions/routes.OpenScraperAttachment'
        type: array
      case_number:
        type: string
      description:
        type: string
      extra_metadata:
        additionalProperties: true
        type: object
      filed_date:
        type: string
      filing_type:
        type: string
      party_name:
        type: string
    type: object
  tasks.AttachmentChildInfo:
    properties:
      extension:
        type: string
      lang:
        type: string
      mdata:
        additionalProperties: {}
        type: object
      name:
        type: string
      url:
        type: string
    type: object
  tasks.CaseInfoMinimal:
    properties:
      case_name:
        type: string
      case_number:
        type: string
      case_type:
        type: string
      case_url:
        type: string
      closed_date:
        description: An RFC3339 DateTime
        type: string
      description:
        type: string
      extra_metadata:
        additionalProperties: true
        type: object
      hearing_officer:
        type: string
      indexed_at:
        description: An RFC3339 DateTime
        type: string
      industry:
        type: string
      opened_date:
        description: An RFC3339 DateTime
        type: string
      petitioner:
        type: string
    type: object
  tasks.CaseInfoPayload:
    properties:
      case_name:
        type: string
      case_number:
        type: string
      case_type:
        type: string
      case_url:
        type: string
      closed_date:
        description: An RFC3339 DateTime
        type: string
      description:
        type: string
      extra_metadata:
        additionalProperties: true
        type: object
      filings:
        items:
          $ref: '#/definitions/tasks.FilingChildInfo'
        type: array
      hearing_officer:
        type: string
      indexed_at:
        description: An RFC3339 DateTime
        type: string
      industry:
        type: string
      opened_date:
        description: An RFC3339 DateTime
        type: string
      petitioner:
        type: string
    type: object
  tasks.FilingChildInfo:
    properties:
      attachments:
        items:
          $ref: '#/definitions/tasks.AttachmentChildInfo'
        type: array
      description:
        type: string
      extra_metadata:
        additionalProperties: true
        type: object
      filed_date:
        description: An RFC3339 DateTime
        type: string
      filing_type:
        type: string
      name:
        type: string
      party_name:
        type: string
    type: object
  tasks.FilingInfoPayload:
    properties:
      case_info:
        $ref: '#/definitions/tasks.CaseInfoMinimal'
      filing:
        $ref: '#/definitions/tasks.FilingChildInfo'
    type: object
  tasks.KesslerTaskInfo:
    properties:
      queue:
        type: string
      state:
        type: string
      status:
        type: string
      task_id:
        type: string
    type: object
  tasks.NYPUCDocInfo:
    properties:
      date_filed:
        type: string
      docket_id:
        type: string
      file_name:
        type: string
      item_no:
        type: string
      name:
        type: string
      nypuc_doctype:
        type: string
      organization:
        type: string
      serial:
        type: string
      url:
        type: string
    type: object
host: api.kessler.xyz
info:
  contact:
    email: support@kessler.xyz
    name: Support
    url: https://kessler.xyz
  description: Document Ingest Service of Dubious Quality
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kessler Ingest API
  version: "1.0"
paths:
  /add-task/ingest:
    post:
      consumes:
      - application/json
      description: Creates a new default ingestion task. This should use the default
        openscrapers schema.
      parameters:
      - description: Filing + Case information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tasks.FilingInfoPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.KesslerTaskInfo'
        "400":
          description: Error decoding request body
          schema:
            type: string
        "500":
          description: Error adding task
          schema:
            type: string
      summary: Add Default Ingest Task
      tags:
      - tasks
  /add-task/ingest/case:
    post:
      consumes:
      - application/json
      description: Creates a new Case-specific ingestion task
      parameters:
      - description: Case information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tasks.CaseInfoPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.KesslerTaskInfo'
        "400":
          description: Error decoding request body
          schema:
            type: string
        "500":
          description: Error adding task
          schema:
            type: string
      summary: Add Case Ingest Task
      tags:
      - tasks
  /add-task/ingest/nypuc:
    post:
      consumes:
      - application/json
      description: Creates a new NYPUC-specific ingestion task
      parameters:
      - description: NYPUC document information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tasks.NYPUCDocInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.KesslerTaskInfo'
        "400":
          description: Error decoding request body
          schema:
            type: string
        "500":
          description: Error adding task
          schema:
            type: string
      summary: Add NYPUC Ingest Task
      tags:
      - tasks
  /add-task/ingest/openscraper:
    post:
      consumes:
      - application/json
      description: Creates a new OpenScraper-specific ingestion task
      parameters:
      - description: OpenScraper filing information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.OpenScraperFiling'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.KesslerTaskInfo'
        "400":
          description: Error decoding request body
          schema:
            type: string
        "500":
          description: Error adding task
          schema:
            type: string
      summary: Add OpenScraper Ingest Task
      tags:
      - tasks
  /task/{id}:
    get:
      description: Retrieves information about a specific task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.KesslerTaskInfo'
        "404":
          description: Error retrieving task info
          schema:
            type: string
        "501":
          description: Not implemented
          schema:
            type: string
      summary: Get Task Information
      tags:
      - tasks
  /version_hash:
    get:
      description: Returns the current version hash of the application
      produces:
      - text/plain
      responses:
        "200":
          description: Version hash string
          schema:
            type: string
      summary: Get Version Hash
      tags:
      - system
swagger: "2.0"